requirements.txt
beautifulsoup4==4.13.1
fpdf==1.7.2
protobuf==5.29.3
python-dotenv==1.0.1
python_docx==1.1.2
Requests==2.32.3
SQLAlchemy==2.0.37
streamlit==1.41.1
trafilatura==2.0.0
validators==0.34.0
google-generativeai
